Perform writeback after an abort in STM instructions. This fixes the endless loop of data aborts!
Per the ARM610 manual:

Aborts during STM instructions

If the abort occurs during a store multiple instruction, ARM610 takes
little action until the instruction completes, whereupon it enters the
data abort trap. The memory manager is responsible for preventing
erroneous writes to the memory. The only change to the internal state of
the processor will be the modification of the base register if write-back
was specified, and this must be reversed by software (and the cause of the
abort resolved) before the instruction may be retried.