Be consistent in handling the arm clock after coprocessor opcodes.
MCR/MRC weren't clocking the arm in opcodes -- instead the
exec code (e.g. inside copro15) was doing this on success.

FPA however was not doing this, so if there were an MCR/MRC
to copro #1, the PC would never change.

So, have FPA do this, which means the STC/LDC opcodes should
not.