Fix horizontally mirroring of the LCD when booting into NewtonOS.
The rationale for believing this reg (0x94) controls this is,
NewtonOS is the only thing seen so far that sets it to 0x06.

Diagnostics mode sets it to 0x05.

And if you crash NewtonOS causing it to print the registers to
the screen, it too will set it back to 0x05 before drawing.

Easy way to do the above:
newton_set_newt_tests(_newton, 0xffffffff);

Notepad will showing the newton tour for a few seconds before
crashing and dumping the regs to the screen.