When Newton reads our ADCSource register, return the last value it wrote, not our shifted value.
This change alters the Nicad battery routine, which required
renumbering it to 0x14.

Rationale behind this is when the J1 ROM is at the tablet alignment
screen, it listens to a few unknown interrupts.  When firing irq18,
it would spew out:

DebugStr: Who screwed up the charge on the tablet?

This was based on it reading the ADCSource, and doing roughly

if (((adcSource << 20) >> 20) != 0x140) {
DebugStr(...);
}

(See the routine at 0x00186414 in the J1 ROM)