Add an ivar, "supportsRegularFiles".
If true, the TapFileCntl will open any file it's asked to.

If false, it'll only open files starting with % (which are
presumed to be tty windows).

The J1 ROM will abort if a regular file is opened, complaining

_sys_flen() not implemented

The Notepad, 1.0 and 1.10 images don't complain this way. And
if we let them open "bootNoCalibrate", then we can skip the
tablet calibration screen -- meaning they progress into
NewtonOS now!