Claim that irq18 belongs to the tablet.
Rationale:  The Lindy diags "Tablet Test" enables it.
Without the interrupt, one cannot progress through the
tablet test. With it, it goes through all the cycles.

So always raise it during touch down, and raise it
during touch up. And for the older ADC based method
which both Junior+Lindy diags wants, only raise the
interrupt if the ADC source is the tablet.

Additionally when firing the IRQ on a tablet alignment screen
in the J1 ROM, prior to earlier ADC fixes, it would spew out:

DebugStr: Who screwed up the charge on the tablet?