Lots of unknown writes to 0xa8, 0xb0, 0xb4, 0xc0, 0xc4, 0xc8 during NewtonOS boot, about the same time one would hear a chime.
Firing IRQ19 at inopportune times has lead to a message:

***************** ACK. sound missed an interrupt

So make an assumption that IRQ19 relates to the sound DMA,
and that these unknown regs relate to sound DMA.

0xb0 seems like the RAM to start DMA from.

Perhaps 0xb4 is the length?

0xb4 is written right after IRQ18 is enabled, so seems
like a perfect place to fire the IRQ to make the Newton
think DMAs are completing.

After it fires 0xb0 is rewritten, often with a new RAM
address.