Begin Sharp LCD driver rewrite...
The Sharp driver exhibits a number of problems, primarily
when booting into NewtonOS, but there were a number of hacks
to get Diagnostics to look correct.

Most of the registers were guesses, and circumstantial.

Thankfully the Junior diagnostics does tell us what the
registers are, via:

Manufacturing -> LCD Unit Check -> LCD Register Check

Prior to this change most of the registers failed, but
with this change all but two pass, helping to validate
these changes.

Armed with this knowledge, the Sharp LCD has 32 registers,
spaced 4 bytes apart. The 32 registers are repeated in our
256 bytes of address space.  If the high bit is not set
(e.g. < 0x80), then it is treated as a read.  If the
high bit is set (>= 0x80), then it is treated as a write.

Clean up our enum of register names, implement all of the
cursor+window registers properly.

The pixel writing code is roughly the same as before to prevent
any regressions (e.g. NewtonOS+diags still boots and is as usable
as it was before).

This commit serves as a good point between the register cleanup,
and the pixel writing(and reading) rewrite.