Remove inverse mode hack used primarily in diagnostics when tapping on items.
Drawing the full 8px would draw outside of the box, so the
hack was introduced to only write half the pixels.  This mostly
worked and the inverse looked okay.

The proper way, this commit, would be to clip the cursor to
the window's coordinates.  If the cursor is outside the window,
we don't copy the bit.

Now the inverse looks perfect (no single lines that don't get
inverted, etc).