When the RTC is written, reset our bootTime variable to the current time.
NewtonOS 2.0's setup wants to set the current date+time.  It would
hang on changing the time, as it would set the RTC to a value,
read the RTC back, and get value + <existing boot time>, which was
too far in the future.  And it would keep looping until this returned
it's expected value, which it would never do.

With the bootTime reset to the current time, the result for the RTC
is effectively the base value Newton just set, plus 0 -- making
NewtonOS happy, and able to progress through the setup.

And now MP130 v2.0 gets out of setup and is usable!