Get the tablet ADC values closer to Apple's default calibration settings.
The X+Y position are only needed for two ADC sources.  The
other two are still a mystery, but they can be constants.

Diags is happy as long as that the sum of four readings of
a source, right shifted by 2, is greater than 512.

NewtonOS alignment is not happy unless the values produced
by the two sources differ by at least 0x03e8.

Prior to this change, if one skipped calibration (via
"bootNoCalibrate" tapfilecntl support), the region where
the tablet would work was limited to a small box in the top-left
corner of the screen (makes sense given it's the LCD's actual
origin).

With this change, the tablet works pretty much everywhere on the
screen, but there is a slight offset from the mouse cursor and
where the point shows up.