Check for bad ROM sizes -- the dump I made of an OMP running v1.3 is missing a byte, resulting in a file size that is 4194303 instead of 4194304 bytes...
This is why the ROM was crashing.  Some of the words towards the
very end of the file were shifted left by a byte, and there's
a list of addresses towards the end. e.g.:

003e6b38         dd         0x1acb9c00
003e6b3c         dd         0x00000000
003e6b40         dd         0x1a1f1400
003e6b44         dd         0x00000200
003e6b48         dd         0x1a1f1800
003e6b4c         dd         0x00000200
003e6b50         dd         0x1a1f1c00
003e6b54         dd         0x00000200
003e6b58         dd         0x1a1f2400
003e6b5c         dd         0x00000300
003e6b60         dd         0x1a1f2c00

All of those addresses should be 0x001a....

The image was encountering a data abort when it'd try
to load a byte from 0x1A1F1400 (instead of 0x001A1F14)
the data abort would end up throwing a evt.ex.abt.bus,
that wasn't caught, so it'd show the system error dialog.

(A 0x00 can be inserted around 0x003e6060 to fix this)