Rename IRQs 25 + 26.
Now that NewtonOS is running, it's easier to see what the behavior
of these are.

IRQ25, now RuntInterruptPowerFault, will cause the Newton to
reboot.  After NewtonOS comes back up, it displays an error whose
wording varies on OS version.  With the Notepad ROM:

Sorry, a problem has occurred. (The system restarted because a PCMCIA
card or serial device was drawing too much power.)

IRQ26, now RuntInterruptBatteryRemoved, causes early NewtonOS
versions to reboot.  After NewtonOS comes back up, it displays
an error similar to:

Sorry, a problem has occurred. (The system restarted because the battery
was dead.)