Handle powering down of the ARM and the Runt.
Newton may ask for the ARM to be paused -- no instructions should
be executed, but the Runt will still be alive -- ticking, firing
interrupts, etc.

Newton may ask for the device to be put to sleep, where both
the ARM and Runt are paused. Like above, no instructions will be
executed, but the Runt will also stop ticking, etc.  The only
thing that will resume it is an external interrupt, e.g. the power
switch.

All of the ROM images except for the Notepad + ARMistice will
attempt to put the device to sleep after a set interval (defined
in Prefs -> Sleep).  Without this change they'd just angrily loop,
and were hard to recover.

With this change, they will go to sleep and the screen will dim
to reflect this.  And the device can be woken back up via
Control-3, or Hardware menu->Power Switch.

(This is all a bit crude, and it'd probably be nice to be able
to exit newton_emulate() entirely, but it suffices as a first pass)