Properly handle orientation/display-mode on the Squirt.
Register 0x70 has nothing to do with it -- this was just
coincidental, enough to work for Lindy Diags (which draws in
landscape) and NewtonOS in portrait.

With Newton 2.0, you can rotate the screen, and things no
longer displayed correctly.

0x54 appears to be the correct register. With this, Lindy Diags
looks fine, NewtonOS 1.x looks fine, and NewtonOS 2.0 in both
landscape + portrait look fine.