Improve the TapFileCntl read function.
Having forgotten about since I wrote it for Einstein...

0 means a call is successful.
len means a call failed and is EOF.
Anything else means it was partially successful.

Per:
http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.dui0376d/Baceahbd.html

Our do_sys_read() returns the length of the read, which doesn't
align with these result codes.

So if do_sys_read() returns -1, return len.

If do_sys_read() returns 0, lie about this, change the buffer to
a length of one with contents 0x00.

And then provided the call didn't fail, return maxlen - length,
meaning we'd only return 0 if we filled the buffer.