Create a FileStream on launch, and a read-only ListenerWindow that'll display the data from this file stream.
Give the newton_t* this file stream as it's log source.

So now any logging from the Newton will go to a Cocoa window,
instead of stdout.

Look in NSUserDefaults for logging flags for the newton+runt,
so that these can be changed via command line defaults write...

Add a menu item under Help to reveal the console window.

If newton_load_rom() fails, display an error message and
don't attempt to emulate anything.