For 8MB ROMs, don't repeat the ROM twice in the 16MB space.  Instead, repeat the last 4MB after the full 8MB.
MP130 v1.2 diagnostics wants to read a lot from
addresses like 0x00abe0ac. With the default mirroring,
this would point back at 0x002bdfbc, which isn't part of the
diagnostics region.  With this mapping, it ends up
pointing at 0x006be0ac, which is in diagnostics region
and happens to be the string "DIAGNOSTICS".
(Without this, diagnostics didn't work. With it, it does)