Improve the tablet analog values.
Bypassing calibration via "bootNoCalibrate", this is really really
close for all of the ROMs except for the Notepad image (and the J1,
since "bootNoCalibrate" doesn't work with it).

Regardless of if "bootNoCalibrate" is ever used, this helps a bit
with the MP130 v2.0 image, as you used to have to click slightly
above the Continue button (or tap the screen a few dozen times)
before it'd take you to the calibration/alignment screen.