Clean up the ADC sample sources.
Inside NewtonOS, when long pressing on the clock to reveal the
temperature, the ADC sample source was being set to 0x00000202.

Due to diags, we know the thermistor to be behind 0x00003202.

The first 12 bits of each are 0x202.

All of the other unknown ADC sources matched up similarly (e.g.
UnknownB was 0x802, and BackupBattery was 0x3802).

This lines up with behavior seen with the tablet sources.
Had seen 0x32, 0x1032, 0x2032, 0x3032, etc.

I'd suspect the last 4 bits are a sampling interval?  The clock
"easter egg" just wanted a one off -- it supplied 0.  Diags
does update the screen regularly, and it supplied 3.