Get Newton to recognize SRAM cards.
This is disabled by default, as the "Format Card?" every launch
is annoying, and it complains (sometimes frequently) about
replacing the low battery in the card.

When the PCMCIA IRQ is raised:

1) Reg 0x6c is read.  Provided it returns a non-zero value:
2) Reg 0x74 is read.
3) Reg 0x64 is written to.

Reg 0x74 is written to generally before PCMCIA IRQs are
enabled on the Runt.  So make the assumption this is an
enabled interrupts register.

Given 0x6c is the first register read after the IRQ
fires, assume this is the active interrupts register.

Given 0x64 is written to after the others are read,
assume this is a clear interrupt register.

Meaning, an interrupt fires, it figures out which one,
which are enabled, and masks it to just clear the fired
ones.

If we fire interrupts when reg 0x74 is written to
with a non-zero value, the system appears to be happy. And
with the next changes, NewtonOS will automatically "mount"
the card after doing stylus alignment.

Now for sketchier territory:

0x7c definitely seems to be a status register (as previously
noted).  In order to pass some diagnostics checks, I'm
reading reg 0x58 to determine what to do.  This could be
purely coincidental...

With that said, if reg 0x58 is 0x00 or 0x0a, a read to 0x7c
seems to return 2 if there's a card present, 0 if not.
This holds true for diagnostics and NewtonOS.

Returning 2 here, and firing the IRQ, causes NewtonOS
to say "The card you inserted cannot be read. Do you
want to complete erase this card?".

If reg 0x58 is 0x0b or 0x1b, then for diagnostics it
seems most interested in VPP flags. There's a slight problem here,
in that the result to pass diagnostics VPP tests
causes other systems to think the inserted card is write
protected. Bit 6 in the result seems to indicate write
protection or not.

If we return a value with bit 6 set, then Newton will
complain "This memory card could not be formatted. (This memory
card is write protected.)".  So while it fails diagnostic
VPP tests, don't include (!vpp2 << 5) in our result.

With that, the SRAM card is formatted, can be backed up to, etc.

And for some output from the J1 image:

??Parser error -10533
$$Read back 5a, a5
$$Recognized card: TCardHandlerMemory::  250 ns 1024K bytes
$$Send Client 	kNewCard 'sram' 14FB 0 off=0 size=100000 BadCIS=1
:-)   TStore ==> kStoreAwaitRegisterReply
:-)   registering with card server
--waking up for stor
:-)   TStore ==> kStoreRegistered