Attempt to make sense of the serial registers on the Runt.
In Lindy diagnostics, when doing Evalutation -> Serial Test,
the serial config register is written to 30 times in a row.

Changing some of the configuration variables (baud, parity, stop bits)
results in only one value in the batch of 30 changing.

So make the assumption that it's a two staged write, the first
being the "internal" serial register, and the second being the
value for this register.

This tends to work -- all of the first writes are small values
(< 0x10) while the second writes vary.

In this setup, it seems like "internal" serial register
0x0c relates to slow speed bauds (<= 57600), 0x04 relates
to parity + stop bits, 0x08 relates to flush (same Lindy
diags writes to it after every byte written to the serial tx data
reg).

For fast speed bauds, it starts talking to a number of runt serial
regs that have never been encountered before.

And make the assumption that the "SerialA" IRQ is actually for
the Infrared, as after firing it, IR registers tend to be read.