Allow membank's to handle byte requests themselves, instead of always doing it for them (via word reads, shifts, word writes, etc)
Optional, so if a membank doesn't supply a get/set uint8 handler,
the old behavior will be used.

Have the memory_t (used for ram, rom, etc) supply these handlers.

(This is to help prepare with serial stuff on the Runt, where
it seems like it needs to keep state, and the set_mem8 behavior
is tripping that state up)