Use byte access on the Runt, which is seemingly only used for the LCD+serial subsystems.
LCD has been observed well enough to know I've never seen a word
access on it.  So abort() those code paths. And swap all the LCD
code to use bytes instead of words, saving on a lot of shifts...

Serial seems to be mostly byte driven, except when the mysterious
base is written to (0x01408000). Don't abort() those paths yet.