Keep more state for the serial ports.
Given it appears that many serial registers hide behind
runt registers 0x800b(ir) and 0x810b(serial), and that
it operates by first storing to this register an "internal"
register it wants to access, then doing a read/write against
this register to retrieve the internal reg's value... we can't
rely on our normal behavior of just returning the last written
value.

So keep track of what the current loaded register value is,
end a bank of registers, and operate on those.

To visualize with the logging enabled:

[SERIAL:CONFIG:LOADREG] reg:0x00
[SERIAL:CONFIG:GETVAL] reg:0x00 => 0xff
[SERIAL:CONFIG:LOADREG] reg:0x09
[SERIAL:CONFIG:SETVAL] reg:0x09 => 0x80
[SERIAL:CONFIG:LOADREG] reg:0x04
[SERIAL:CONFIG:SETVAL] reg:0x04 => 0x44