Fix regression for 8MB ROM mirroring.
Passing the absolute address to the membank handler killed
the "trick" of adding two mappings for the 8MB ROM as the
underlying memory_t didn't know of the two, and just subtracted
it's base from the address.

So, give memory_t's the concept of mappings from virtual to
physical addresses.  This way we can tell it to map
a virtual address 0x00800000 to the physical address
0x00400000, for 4MB, to achieve our desired goal.