The Zilog Z85C30 datasheet states there are write registers, and there are read registers, and they don't always correspond to the same thing.
This behavior seems correct for the Newton given some observed
behavior:

When we raise a Runt Serial IRQ, it tends to go read reg 3 on
channel A (IR). This would correspond to the read register
interrupt pending bits.

It subsequently reads reg 2, which would be the interrupt vectors.

It subsequently writes to reg 9, the "Master interrupt control
and reset", resetting channel B (serial) and enabling MIE.

Read register 0 is read from a lot, which wouldn't correspond
to the write register's purpose "CRC initialize, ..."
If we don't include RR0_TXEMPTY in the result, then it tends
to angrily loop a lot until we do.

Similar for read register 1 -- if RR1_ALLSENT isn't in the
result, it'll angrily loop.

For Lindy diags, in the Evaluation -> Serial Test, this
seems to be a loop-back test of sorts.

After writing a byte, it wants to see RR0_RXAVAIL in reg 0.
When that's set, it then reads from register 8. If
the contents aren't what it just wrote, it'll stop after
~ 8 writes, and fail with "LOW SPEED ASYNC FAIL".

If we do set RR0_RXAVAIL, and put that last written
byte back into read-reg 8, it'll do 256 writes, and then
fail with "RTSB/TXE"

(This is the purpose of the #define SERIAL_LOOPBACK)