Try to handle the modified interrupt vector (read reg 2 from channel B).
Rename irChannel and serialChannel to channelB and channelA.

With these, the Notepad 1.0b1 image will write 40 bytes
out to the serial port during Connection (Computer, really).
It then starts trying to read register #40, which probably
indicates a failure in the regLoaded logic.

The other 1.x images write a byte, before also trying to
interact with this reg #40.