Remove enabledSerialInterrupts.  Just write the value to both channels' registers[2]
If we see a load-reg with 0x28, treat that as a clear pending IRQ
flag.

When the ROM does not know how to handle the IRQ, it will write
to reg9 (Master interrupt control) and clear the pending IRQs that
way.

When the ROM *does* know how to handle the IRQ, it never hits
that reg9 codepath.  Instead after processing the IRQ, they all
seem to "load reg" 0x28.  And then they'd angrily loop, seeing
the pending IRQs haven't changed.

If we clear the pending IRQs on this mysterious write, they
stop with the angry looping.

All ROMs now just intermittently write an 0x30 out, handle the
TX IRQ, sleep, loop.

For all ROMs except for the Notepad (who wasn't affected
by this before), the "Stop" button in the Connection dialog now works!

For the Notepad ROM, the text in the dialog now changes.
"Waiting for a response..." and then "Preparing to connect..."
It previously wrote out a string of different bytes, but now it
too just pings out with a 0x30.