Be the serial output handler for the device's z85c30 chip.
If we're booted into diagnostics mode, automatically enable
a serial loop back mode.

If we're booted into normal mode, feed the serial output into
our docker.  If it's able to make sense of it and generate
a response, send that back into the serial port.

With this, one can hit Connection in the Extras drawer, and
the dialog will almost immediately disappear. Instead of
previously waiting to timeout and displaying an error -- proof
that the docker is communicating properly, and it's issuing a
disconnect command.