Send LA responses when uploading package bytes.
This has been tested on a real Newton and it does work -- there
is an issue after the package is uploaded that the disconnect
doesn't seem to work, the dialog on the Newton remains until
it times out.  But the package is present afterwards.

This isn't working on Leibniz.  All of the bytes of the package
are read in, then the Newton immediately LD's the connection
and displays an error dialog (-10401).  Perplexing...